+++ R/initialization.R
@@ -164,11 +164,10 @@ prior_sample <- function(dat, model = "drm") {
   } else {
     out <-
       c(out,
-        list(phi = array(exp(
-                 dat$pr_phi_mu +
-                 dat$pr_phi_sd * stats::rt(1, df = 3)
-             ),
-             dim = 1)))
+        list(phi = array(rgamma(1,
+                                shape = dat$pr_phi_a,
+                                rate = dat$pr_phi_b),
+                         dim = 1)))
   }
   if (dat$time_ar) {
     out <-